Moved DB.Element extension methods to External.DB.Extensions namespace.
