+++ src/RhinoInside.Revit/Extensions/RevitAPI.cs
@@ -10,25 +10,6 @@ namespace RhinoInside.Revit
 {
   /*internal*/ public static class RevitAPI
   {
-    #region Geometry
-    public static GeometryElement GetGeometry(this Element element, ViewDetailLevel viewDetailLevel, out Options options)
-    {
-      options = new Options { ComputeReferences = true, DetailLevel = viewDetailLevel };
-      var geometry = element.get_Geometry(options);
-
-      if (!(geometry?.Any() ?? false) && element is GenericForm form && !form.Combinations.IsEmpty)
-      {
-        geometry.Dispose();
-
-        options.IncludeNonVisibleObjects = true;
-        return element.get_Geometry(options);
-      }
-
-      return geometry;
-    }
-
-    #endregion
-
     #region ElementId
     public static bool IsValid(this ElementId id)     => id is object && id != ElementId.InvalidElementId;
     public static bool IsBuiltInId(this ElementId id) => id is object && id <= ElementId.InvalidElementId;
@@ -362,30 +343,7 @@ namespace RhinoInside.Revit
     }
     #endregion
 
-    #region Element
-#if !REVIT_2019
-    public static IList<ElementId> GetDependentElements(this Element element, ElementFilter filter)
-    {
-      try
-      {
-        // Start a dry transaction that will be rolled back later
-        using (var transaction = new Transaction(element.Document, nameof(GetDependentElements)))
-        {
-          transaction.Start();
-
-          var collection = element.Document.Delete(element.Id);
-          if (filter is null)
-            return collection?.ToList();
-
-          return collection?.Where(x => filter.PassesFilter(element.Document, x)).ToList();
-        }
-      }
-      catch { }
-
-      return default;
-    }
-#endif
-
+    #region Instance
     public static void SetTransform(this Instance element, XYZ newOrigin, XYZ newBasisX, XYZ newBasisY)
     {
       var current = element.GetTransform();
